Refactor window struct and loops to use uint32_t types
Replaced int types with uint32_t for cursor_pos and num_toolbar_buttons in window_t and related code. Updated loop variables and message struct fields for consistency and type safety. Minor cleanup in strnwidth function.

+++ commctl/edit.c
@@ -84,7 +84,7 @@ result_t win_textedit(window_t *win, uint32_t msg, uint32_t wparam, void *lparam
           }
           break;
         case SDL_SCANCODE_RIGHT:
-          if (win->cursor_pos < (int)strlen(win->title) && win->editing) {
+          if (win->cursor_pos < strlen(win->title) && win->editing) {
             win->cursor_pos++;
           }
           break;
